requirements.txt
streamlit
pandas
geopandas
shapely
python-docx
fpdf
python-dotenv
folium
streamlit-folium
openpyxl
